Pause Menu (in progress) + few other files
Main character and a few rooms added, final sol counter implemented (the Sol 1 to 4 scenes are probably useless now), final pause menu in progress (some Ctrl-C + Ctrl-V required with the code).

Also the airlock (if it's the one that will be used in the final version (if there's a final version)) is ready to get a save feature (in the middle of the scene there's a spot where when you walk in, you can save the current game).