Begining of the end
The original (prototype of the) main title will be deleted after this version, so I push this to Git just in case.

So far the game is mostly playable, there's just the materials to finish, the "barriers" ahead of each room (requiring materials to be opened) and a few details like that.